requirements.txt
psycopg2==2.5.1
Django==1.5.4
PyYAML==3.10
argparse==1.2.1
dj-database-url==0.2.1
dj-static==0.0.5
django-mptt==0.6.0
django-override-settings==1.2
django-permission==0.4.5
django-social-auth==0.7.28
gunicorn==17.5
httplib2==0.8
mock==1.0.1
oauth2==1.5.211
python-openid==2.2.5
setuptools-git==1.0
static==0.4
wsgiref==0.1.2
PIL==1.1.7
requests==2.0.1
beautifulsoup4==4.3.2
versiontools==1.9.1
django-appconf==0.6
django-endless-pagination==2.0
tornado==2.4.1
sockjs-tornado==1.0.0